* деление на компоненты
